eSetFilter {genefilter}	R Documentation

A function to filter an eSet object

Description

Given a Bioconductor's ExpressionSet object, this function filters genes using a set of selected filters.

Usage


eSetFilter(eSet)
getFilterNames()
getFuncDesc(lib = "genefilter", funcs = getFilterNames())
getRdAsText(lib)
parseDesc(text)
parseArgs(text)
showESet(eSet)
setESetArgs(filter)
isESet(eSet)

Arguments

eSet
eSet an ExpressionSet object

lib
lib a character string for the name of an R library where functions of interests reside

funcs
funcs a vector of character strings for names of functions of interest

text
text a character of string from a filed (e. g. description, argument, ..) filed of an Rd file for a fucntion

filter
filter a character string for the name of a filter function

Details

A set of filters may be selected to filter genes in through each of the filters in the order the filters have been selected

Value

A logical vector of length equal to the number of rows of 'expr'. The values in that vector indicate whether the corresponding row of 'expr' passed the set of filter functions.

Author(s)

Jianhua Zhang

See Also

genefilter

Examples


 if( interactive() ) {
   data(sample.ExpressionSet)      
   res <- eSetFilter(sample.ExpressionSet)
 }

[Package genefilter version 1.66.0 Index]