filter_volcano {genefilter}	R Documentation

Volcano plot for overall variance filtering

Description

Generate a volcano plot contrasting p-value with fold change (on the log scale), in order to visualize the effect of filtering on overall variance and also assign significance via p-value.

Usage


filter_volcano(
               d, p, S,
               n1, n2,
               alpha, S_cutoff,
               cex = 0.5, pch = 19,
               xlab = expression(paste(log[2], " fold change")),
               ylab = expression(paste("-", log[10], " p")),
               cols = c("grey80", "grey50", "black"),
               ltys = c(1, 3),
               use_legend = TRUE,
               ...
               )

Arguments

d
Fold changes, typically on the log scale, base 2.

p
The p-values

S
The overall standard deviation filter statistics, i.e., the square roots of the overall variance filter statistics.

n1
Sample size for group 1.

n2
Sample size for group 2.

alpha
Significance cutoff used for p-values.

S_cutoff
Filter cutoff used for the overall standard deviation in S.

cex
Point size for plotting.

pch
Point character for plotting.

xlab
Label for x-axis.

ylab
Label for y-axis.

cols
A vector of three colors used for plotting. These correspond to filtered data, data which pass the filter but are insignificant, and data pass the filter and are also statistically significant.

ltys
The induced bound on log-scale fold change is plotted, as is the significance cutoff for data passing the filter. The ltys argument gives line styles for these drawing these two thresholds on the plot.

use_legend
Should a legend for point color be produced?

...
Other arguments for plot.

Author(s)

Richard Bourgon <bourgon@ebi.ac.uk>

Examples


# See the vignette: Diagnostic plots for independent filtering

[Package genefilter version 1.66.0 Index]