kOverA {genefilter}	R Documentation

A filter function for k elements larger than A.

Description

kOverA returns a filter function with bindings for k and A. This function evaluates to TRUE if at least k of the arguments elements are larger than A.

Usage


kOverA(k, A=100, na.rm=TRUE)

Arguments

A
The value you want to exceed.

k
The number of elements that have to exceed A.

na.rm
If set to TRUE any NA's will be removed.

Value

A function with bindings for A and k.

Author(s)

R. Gentleman

See Also

pOverA

Examples


   fg <- kOverA(5, 100)
   fg(90:100)
   fg(98:110)

[Package genefilter version 1.66.0 Index]