kappa_p {genefilter}	R Documentation

Compute proportionality constant for fold change bound.

Description

Filtering on overall variance induces a lower bound on fold change. This bound depends on the significance of the evidence against the null hypothesis, an is a multiple of the cutoff used for an overall variance filter. It also depends on sample size in both of the groups being compared. These functions compute the multiplier for the supplied p-values or t-statistics.

Usage


kappa_p(p, n1, n2 = n1)
kappa_t(t, n1, n2 = n1)

Arguments

p
The p-values at which to compute the multiplier.

t
The t-statistics at which to compute the multiplier.

n1
Sample size for class 1.

n2
Sample size for class 2.

Value

A vector of multipliers: one per p-value or t-static in p or t.

Author(s)

Richard Bourgon <bourgon@ebi.ac.uk>

Examples


# See the vignette: Diagnostic plots for independent filtering

[Package genefilter version 1.66.0 Index]