maxA {genefilter}	R Documentation

A filter function to filter according to the maximum.

Description

maxA returns a function with the parameter A bound. The returned function evaluates to TRUE if any element of its argument is larger than A.

Usage


maxA(A=75, na.rm=TRUE)

Arguments

A
The value that at least one element must exceed.

na.rm
If TRUE then NA's are removed.

Value

maxA returns a function with an environment containing a binding for A.

Author(s)

R. Gentleman

See Also

pOverA

Examples


   ff <- maxA(30)
   ff(1:10)
   ff(28:31)

[Package genefilter version 1.66.0 Index]