pOverA {genefilter}	R Documentation

A filter function to filter according to the proportion of elements larger than A.

Description

A function that returns a function with values for A, p and na.rm bound to the specified values. The function takes a single vector, x, as an argument. When the returned function is evaluated it returns TRUE if the proportion of values in x that are larger than A is at least p.

Usage


pOverA(p=0.05, A=100, na.rm=TRUE)

Arguments

A
The value to be exceeded.

p
The proportion that need to exceed A for TRUE to be returned.

na.rm
If TRUE then NA's are removed.

Value

pOverA returns a function with bindings for A, p and na.rm. This function evaluates to TRUE if the proportion of values in x that are larger than A exceeds p.

Author(s)

R. Gentleman

See Also

cv

Examples


  ff<- pOverA(p=.1, 10)
  ff(1:20)
  ff(1:5)

[Package genefilter version 1.66.0 Index]