rowROC-class {genefilter}	R Documentation

Class "rowROC"

Description

A class to model ROC curves and corresponding area under the curve as produced by rowpAUCs.

Objects from the Class

Objects can be created by calls of the form new("rowROC", ...).

Slots

data:

Object of class "matrix" The input data.

ranks:

Object of class "matrix" The ranked input data.

sens:

Object of class "matrix" Matrix of senitivity values for each gene at each cutpoint.

spec:

Object of class "matrix" Matrix of specificity values for each gene at each cutpoint.

pAUC:

Object of class "numeric" The partial area under the curve (integrated from 0 to p.

AUC:

Object of class "numeric" The total area under the curve.

factor:

Object of class "factor" The factor used for classification.

cutpoints:

Object of class "matrix" The values of the cutpoints at which specificity ans sensitivity was calculated. (Note: the data is ranked prior to computation of ROC curves, the cutpoints map to the ranked data.

caseNames:

Object of class "character" The names of the two classification cases.

p:

Object of class "numeric" The limit to which pAUC is integrated.

Methods

show signature(object="rowROC")

Print nice info about the object.

[ signature(x="rowROC", j="missing")

Subset the object according to rows/genes.

plot signature(x="rowROC", y="missing")

Plot the ROC curve of the first row of the object along with the pAUC. To plot the curve for a specific row/gene subsetting should be done first (i.e. plot(rowROC[1]).

pAUC signature(object="rowROC", p="numeric", flip="logical")

Integrate area under the curve from 0 to p. This method returns a new rowROC object.

AUC signature(object="rowROC")

Integrate total area under the curve. This method returns a new rowROC object.

sens signature(object="rowROC")

Accessor method for sensitivity slot.

spec signature(object="rowROC")

Accessor method for specificity slot.

area signature(object="rowROC", total="logical")

Accessor method for pAUC slot.

Author(s)

Florian Hahne <fhahne@fhcrc.org>

References

Pepe MS, Longton G, Anderson GL, Schummer M.: Selecting differentially expressed genes from microarray experiments. Biometrics. 2003 Mar;59(1):133-42.

See Also

rowpAUCs

Examples


library(Biobase)
require(genefilter)
data(sample.ExpressionSet)
roc <- rowpAUCs(sample.ExpressionSet, "sex", p=0.5)
roc
area(roc[1:3])

if(interactive()) {
par(ask=TRUE)
plot(roc)
plot(1-spec(roc[1]), sens(roc[2]))
par(ask=FALSE)
}

pAUC(roc, 0.1)
roc

[Package genefilter version 1.66.0 Index]