rowSds {genefilter}	R Documentation

Row variance and standard deviation of a numeric array

Description

Row variance and standard deviation of a numeric array

Usage


rowVars(x, ...)
rowSds(x, ...)

Arguments

x
An array of two or more dimensions, containing numeric, complex, integer or logical values, or a numeric data frame.

...
Further arguments that get passed on to rowMeans and rowSums.

Details

These are very simple convenience functions, the main work is done in rowMeans and rowSums. See the function definition of rowVars, it is very simple.

Value

A numeric or complex array of suitable size, or a vector if the result is one-dimensional. The ‘dimnames’ (or ‘names’ for a vector result) are taken from the original array.

Author(s)

Wolfgang Huber http://www.ebi.ac.uk/huber

See Also

rowMeans and rowSums

Examples


   a = matrix(rnorm(1e4), nrow=10)
   rowSds(a)

[Package genefilter version 1.66.0 Index]