ttest {genefilter}	R Documentation

A filter function for a t.test

Description

ttest returns a function of one argument with bindings for cov and p. The function, when evaluated, performs a t-test using cov as the covariate. It returns TRUE if the p value for a difference in means is less than p.

Usage


ttest(m, p=0.05, na.rm=TRUE)

Arguments

m
If m is of length one then it is assumed that elements one through m of x will be one group. Otherwise m is presumed to be the same length as x and constitutes the groups.

p
The p-value for the test.

na.rm
If set to TRUE any NA's will be removed.

Details

When the data can be split into two groups (diseased and normal for example) then we often want to select genes on their ability to distinguish those two groups. The t-test is well suited to this and can be used as a filter function.

This helper function creates a t-test (function) for the specified covariate and considers a gene to have passed the filter if the p-value for the gene is less than the prespecified p.

Value

ttest returns a function with bindings for m and p that will perform a t-test.

Author(s)

R. Gentleman

See Also

kOverA, Anova, t.test

Examples


  dat <- c(rep(1,5),rep(2,5))
  set.seed(5)
  y <- rnorm(10)
  af <- ttest(dat, .01)
  af(y)
  af2 <- ttest(5, .01)
  af2(y)
  y[8] <- NA
  af(y)
  af2(y)
  y[1:5] <- y[1:5]+10
  af(y)

[Package genefilter version 1.66.0 Index]